Implement a more structured API response
Following the last implementation (i.e., the main query route), this
commit includes a more structured JSON-encoded response in the API. The
response message is now more standardized, even in exception-handling
cases, and can be created more dinamically.

This "dinamic behaviour" led the need of the refactoration of class
`Response` to `Responder`, due of the possibility of ambiguity between
the object of class `Response` and the query response itself.

As a minor note, two Flask's error handlers were added to the code.